+++ packages/synapse-interface/pages/stake/StakeCard.tsx
@@ -131,7 +131,7 @@ const StakeCard = ({ address, chainId, pool }: StakeCardProps) => {
               : formatBigIntToString(
                   lpTokenBalance,
                   tokenInfo.decimals,
-                  18
+                  6
                 )}{' '}
             <span className="text-base text-[#A9A5AD]">
               {pool ? pool.symbol : ''}
@@ -141,7 +141,7 @@ const StakeCard = ({ address, chainId, pool }: StakeCardProps) => {
         <div className="flex items-center justify-between my-2">
           <div className="text-[#EEEDEF]">Staked</div>
           <div className="text-white ">
-            {formatBigIntToString(userStakeData.amount, tokenInfo.decimals, 18)}{' '}
+            {formatBigIntToString(userStakeData.amount, tokenInfo.decimals, 6)}{' '}
             <span className="text-base text-[#A9A5AD]">
               {pool ? pool.symbol : ''}
             </span>
@@ -154,7 +154,7 @@ const StakeCard = ({ address, chainId, pool }: StakeCardProps) => {
           <div className="text-white ">
             {userStakeData.reward === 0n
               ? '\u2212'
-              : formatBigIntToString(userStakeData.reward, 18, 18)}{' '}
+              : formatBigIntToString(userStakeData.reward, 18, 6)}{' '}
             <span className="text-base text-[#A9A5AD]">
               {pool?.customRewardToken ?? 'SYN'}
             </span>