+++ services/sinner/fetcher/client/types/doc.go
@@ -1,2 +1,0 @@
-// Package types provides the GraphQL types for the Scribe service.
-package types