+++ services/sinner/logger/doc.go
@@ -1,6 +1,0 @@
-// Package logger handles logging various sinner events and errors.
-package logger
-
-// check unparam
-// check cyclop
-// check gocognit,cyclop,maintidx